+++ clash-prelude/src/Clash/HaskellPrelude.hs
@@ -4,8 +4,8 @@
   Maintainer  :  QBayLogic B.V <devops@qbaylogic.com>
 
 "Clash.HaskellPrelude" re-exports most of the Haskell "Prelude" with the exception of
-the following: (++), (!!), concat, drop, foldl, foldl1, foldr, foldr1, head,
-init, iterate, last, length, map, repeat, replicate, reverse, scanl, scanr,
+the following: (++), (!!), concat, drop, even, foldl, foldl1, foldr, foldr1, head,
+init, iterate, last, length, map, odd, repeat, replicate, reverse, scanl, scanr,
 splitAt, tail, take, unzip, unzip3, zip, zip3, zipWith, zipWith3.
 -}
 
@@ -18,6 +18,6 @@ module Clash.HaskellPrelude
 where
 
 import Prelude hiding
-  ((++), (!!), concat, concatMap, drop, foldl, foldl1, foldr, foldr1, head, init,
-   iterate, last, length, map, repeat, replicate, reverse, scanl, scanr, splitAt,
+  ((++), (!!), concat, concatMap, drop, even, foldl, foldl1, foldr, foldr1, head, init,
+   iterate, last, length, map, odd, repeat, replicate, reverse, scanl, scanr, splitAt,
    tail, take, unzip, unzip3, zip, zip3, zipWith, zipWith3, undefined, (^))