+++ tests/shouldwork/Numbers/NumConstantFolding_1.hs
@@ -65,6 +65,21 @@ cIntegral = (rQuot,rRem,rDiv,rMod,rQuotRem,rDivMod)
     rDivMod  = divMod  @n (lit 22155) (lit 46)
     -- toInteger?
 
+cParity :: forall n. (Num n, Parity n) => _
+cParity = (rEven, rOdd)
+  where
+    rEven = even @n (lit 22290)
+    rOdd  = odd  @n (lit 22291)
+
+cParityBS0 :: forall n. (Num n, Parity n, BitPack n, 0 ~ (BitSize n)) => _
+cParityBS0 = (rEven, rOdd)
+  where
+    f :: Int -> Bool -> Int
+    f a = \x -> case x of {True -> 0; False -> a}
+
+    rEven = f 22302 $ even @n (lit 22300)
+    rOdd  = f 22303 $ odd  @n (lit 22301) == False
+
 cBitsNoPopCount :: forall n. (Num n, Bits n) => _
 -- the nested tuple here is so we can still 'show' the result
 cBitsNoPopCount = ((r00,r01,r02,r03,r04,r05,r06,r07,r08,r09,r10),r11,r12,r13,r14,r15,r16,r17,r18,r19)
@@ -196,6 +211,8 @@ tUnsigned16
     , cEq             @(Unsigned 16)
     , cOrd            @(Unsigned 16)
     , cIntegral       @(Unsigned 16)
+    , cParity         @(Unsigned 16)
+    , cParityBS0      @(Unsigned 0 )
     , cBits           @(Unsigned 16)
     -- , cFiniteBits  @(Unsigned 16) -- broken
     , csClashSpecific @(Unsigned 16)
@@ -207,6 +224,8 @@ tSigned16
     , cEq             @(Signed 16)
     , cOrd            @(Signed 16)
     , cIntegral       @(Signed 16)
+    , cParity         @(Signed 16)
+    , cParityBS0      @(Signed 0 )
     , cBits           @(Signed 16)
     -- , cFiniteBits  @(Signed 16) -- broken
     , csClashSpecific @(Signed 16)
@@ -218,6 +237,8 @@ tBitVector16
     , cEq             @(BitVector 16)
     , cOrd            @(BitVector 16)
     , cIntegral       @(BitVector 16)
+    , cParity         @(BitVector 16)
+    , cParityBS0      @(BitVector 0 )
     , cBits           @(BitVector 16)
     -- , cFiniteBits  @(BitVector 16) -- broken
     , csClashSpecific @(BitVector 16)
@@ -229,12 +250,15 @@ tIndex
     , cEq             @(Index 50000)
     , cOrd            @(Index 50000)
     , cIntegral       @(Index 50000)
+    , cParity         @(Index 50000)
+    , cParityBS0      @(Index 1    )
     , cBits           @(Index 65536)
     -- , cFiniteBits     @(Index 50000) -- broken
     , cIndex1 -- ensure special case for index 1 is verified
     , csClashSpecific @(Index 50000)
     , cResize         @(Index 50000)
     )
+
 tSFixed
   = ( cNum            @(SFixed 16 0)
     , cEq             @(SFixed 16 0)